rename errorReportApiClient, update interface docs
