+++ src/main/java/com/bugsnag/android/Client.java
@@ -62,7 +62,7 @@ public class Client extends Observable implements Observer {
     protected final User user = new User();
     protected final ErrorStore errorStore;
     private final EventReceiver eventReceiver = new EventReceiver();
-    private ReportApiClient reportApiClient = new DefaultHttpClient();
+    private ErrorReportApiClient errorReportApiClient = new DefaultHttpClient();
 
     /**
      * Initialize a Bugsnag client
@@ -165,7 +165,7 @@ public class Client extends Observable implements Observer {
         config.addObserver(this);
 
         // Flush any on-disk errors
-        errorStore.flush(reportApiClient);
+        errorStore.flush(errorReportApiClient);
     }
 
     public void notifyBugsnagObservers(NotifyType type) {
@@ -510,11 +510,11 @@ public class Client extends Observable implements Observer {
     }
 
     @SuppressWarnings("ConstantConditions")
-    void setReportApiClient(@NonNull ReportApiClient reportApiClient) {
-        if (reportApiClient == null) {
-            throw new IllegalArgumentException("ReportApiClient cannot be null.");
+    void setErrorReportApiClient(@NonNull ErrorReportApiClient errorReportApiClient) {
+        if (errorReportApiClient == null) {
+            throw new IllegalArgumentException("ErrorReportApiClient cannot be null.");
         }
-        this.reportApiClient = reportApiClient;
+        this.errorReportApiClient = errorReportApiClient;
     }
 
     /**
@@ -808,7 +808,7 @@ public class Client extends Observable implements Observer {
                 break;
             case ASYNC_WITH_CACHE:
                 errorStore.write(error);
-                errorStore.flush(reportApiClient);
+                errorStore.flush(errorReportApiClient);
         }
 
         // Add a breadcrumb for this error occurring
@@ -817,7 +817,7 @@ public class Client extends Observable implements Observer {
 
     void deliver(Report report, Error error) {
         try {
-            reportApiClient.postReport(config.getEndpoint(), report);
+            errorReportApiClient.postReport(config.getEndpoint(), report);
             Logger.info(String.format(Locale.US, "Sent 1 new error to Bugsnag"));
         } catch (DefaultHttpClient.NetworkException e) {
             Logger.info("Could not send error(s) to Bugsnag, saving to disk to send later");
